server/app/security/auth_service.py
import jwt
import os
from datetime import datetime, timedelta, timezone
from dotenv import load_dotenv

load_dotenv()

SECRET_KEY = os.getenv("JWT_SECRET_KEY")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_HOURS = 24

# Validate that SECRET_KEY exists and is secure
if not SECRET_KEY:
    raise ValueError("JWT_SECRET_KEY environment variable is required")
if len(SECRET_KEY) < 32:
    raise ValueError("JWT_SECRET_KEY must be at least 32 characters long")

def create_access_token(user_id: str, email: str) -> str:
    """Create JWT token for authenticated user"""
    if not user_id or not email:
        raise ValueError("user_id and email are required")
        
    now = datetime.now(timezone.utc)
    expire = now + timedelta(hours=ACCESS_TOKEN_EXPIRE_HOURS)
    payload = {
        "user_id": user_id,
        "email": email,
        "exp": expire,
        "iat": now
    }
    return jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)

def verify_token(token: str) -> dict | None:
    """Verify JWT token and return payload"""
    if not token or not isinstance(token, str):
        return None
        
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        # Validate required fields exist
        if not payload.get("user_id") or not payload.get("email"):
            return None
        return payload
    except jwt.ExpiredSignatureError:
        return None
    except jwt.InvalidTokenError:
        return None