server/requirements.txt
fastapi
uvicorn
python-dotenv
bcrypt
motor
requests 
pydantic[email]
pyjwt
pytest
pytest-asyncio
